add holding tab for concentrator
